==> notebooks/research/1b.run_defaults.ipynb <==
# Run default configurations


Because the default hyperparameter settings are not necessarily included in the grid, we run these seperately.  
These config files were made manually. See `config/defaults_all_v2.toml` for all the used parameters.  
*note: the experiments are ran for way more datasets than actually used in the final results*

%load_ext autoreload
%autoreload 2

## Import and config

import numpy as np
from ODD.workflows.experiment_config import execute_experiment, get_results_from_config
from pathlib import Path

VERSION=2

config_name = Path()/'config'/f'defaults_all_v{VERSION}.toml'
result_name = Path()/'results'/f'defaults_all_v{VERSION}.pkl'
processed_dir = Path()/'processed_results_v5'

scheduler = "domain_to_dask_scheduler.com:8786"

## Run the algorithms based on the config file

result_df = execute_experiment(config_name, result_name, dask_scheduler = scheduler, shuffle = True, progress = False)
result_df = get_results_from_config(config_name)

## Directly write these results to the processed results directory
(because there is no further processing of these results necessary)

for algo_name, df in result_df.groupby('algo_name'):
    # rename IF to IForest
    if algo_name == 'IF': 
        algo_name = 'IForest'
    processed_name = processed_dir/f"default_performance"/ f"{algo_name}.csv"
    processed_name.parent.mkdir(parents = True, exist_ok = True)
    df.to_csv(processed_name)
    processed_name = processed_dir/f"default_performance"/ f"{algo_name}.pkl"
    df.to_pickle(processed_name)
