==> notebooks/research/3a.final_result_table.ipynb <==
import pandas as pd 
import numpy as np
from pathlib import Path
from ODD.analysis.dataset_selection import get_datasets_to_use
from ODD.analysis.hyperparameters import select_peak_performance, select_best_average_performance, calculate_best_average_performance, calculate_validation_set_performances
from ODD.analysis.result_processing import average_performance_per_method, average_aligned_ranks_with_versions, average_ranks_with_versions, average_ranks_with_versions_and_nemenyi
from ODD.analysis.result_analysis_charts import *
from tqdm import tqdm
import altair as alt 
alt.data_transformers.disable_max_rows()
# alt.renderers.enable('png')
from collections import defaultdict


%load_ext autoreload
%autoreload 2

# Config

VERSION = 1 
algorithms = ['CBLOF', 'HBOS', 'IForest', 'KNN' , 'LOF', 'OCSVM']
grid_versions_to_use = defaultdict(lambda: 1)
grid_versions_to_use['HBOS'] = 2
grid_versions_to_use['CBLOF'] = 2
grid_versions_to_use['IForest'] = 2 

result_path = Path()/'results'
processed_path = Path()/'processed_results_v4'

# Single function to generate a result table

def compare_algorithms_under_evaluation_setting(evaluation_setting_path, datasets_to_use, alpha, validation_reference, validation_df = None):
    # read the correct results 
#     try: 
    print(evaluation_setting_path)
    relevant_dataframes = [pd.read_pickle(evaluation_setting_path / f"{algo}.pkl") for algo in algorithms]
#     except: 
#         relevant_dataframes = [pd.read_csv(evaluation_setting_path / f"{algo}.csv") for algo in algorithms]
        
    all_results = pd.concat(relevant_dataframes, axis = 0, keys = algorithms, names = ['algorithm_name']).droplevel(1).reset_index()
    
    # only keep the results for the correct datasets 
    filtered_results= all_results.groupby(['dataset_id', 'anomaly_fraction']).filter(lambda x: x.name in datasets_to_use.index)
    
    # drop all columns that might cause problems 
    filtered_results = filtered_results.drop(columns = ['auc', 'ap', 'validation_indices', 'test_auc', 'test_ap'], errors = 'ignore')
    
    # calculate performance on test set 
    filtered_results = calculate_validation_set_performances(filtered_results, validation_df)
        
    # rename test_auc to auc 
    filtered_results = filtered_results.rename(columns = {'test_auc': 'auc', 'test_ap':'ap'})
    
    filtered_results = filtered_results[['algorithm_name', 'dataset_id', 'auc', 'ap']]
    filtered_results = filtered_results.rename(columns = {'algorithm_name':'algo_name'})
    
    avg_performance = average_performance_per_method(filtered_results)
#     avg_aligned_ranks = average_aligned_ranks_with_versions(filtered_results)
    avg_ranks, critical_distance = average_ranks_with_versions_and_nemenyi(filtered_results, alpha)
    performance_df = pd.concat([avg_performance, avg_ranks], axis = 1).reindex(columns = ['avg_auc', 'auc_ranks', 'avg_ap', 'ap_ranks'])
    return performance_df, critical_distance

def generate_result_table(
    evaluation_settings,  
    reference_for_validation_set = None, 
    datasets_to_use = None,
    alpha = 0.05,
):  
    print(f'Doing comparison based on {len(datasets_to_use)} datasets')
    
    # calculate the best average performance if necessary (could be simplified now that datasets are fixed)
    if 'best_average_performance' in evaluation_settings: 
        print('calculating best average performance')
        calculate_best_average_performance(Path()/'results', algorithms, grid_versions_to_use, datasets_to_use, processed_path/'best_average_performance')
    
    # read the validation sets to use
    validation_df = None
    if reference_for_validation_set is not None: 
        try: 
            validation_df = pd.read_pickle(processed_path/reference_for_validation_set/'reference_validation.pkl')
        except: 
            validation_df = pd.read_pickle(processed_path/reference_for_validation_set/'validation_set.pkl')
    
    # do all the comparisons
    all_comparisons = [
        compare_algorithms_under_evaluation_setting(
            processed_path/evaluation_setting, 
            datasets_to_use, 
            alpha, 
            validation_reference = reference_for_validation_set,
            validation_df = validation_df) 
        for evaluation_setting in evaluation_settings]
    
    # make them into a single result table 
    all_result_table = pd.concat([tup[0] for tup in all_comparisons], axis = 1, keys = evaluation_settings)
    
    # critical distances should be the same so just pick one 
    critical_distances = all_comparisons[0][1]
    assert all(comparison[1] == critical_distances for comparison in all_comparisons)
    
    return all_result_table, critical_distances

def to_latex(result_table, critical_dist = None, index = True, opmaak = True):
    table = result_table.round(2).astype('str')
    def opmaak_f(column):
        if 'algo' in column.name[1]: 
            return column
        numbers = column.astype(float)
        if 'avg' in column.name[1]:
            max_value = numbers.max()
            best_algorithms = column.index[numbers == max_value]
            column[best_algorithms] = "\\tu{"+column[best_algorithms] + '}'
            return column
        else:
    #         critical_diff = differences.squeeze()[column.name[0]]

            max_value = numbers.min()
            best_algorithms = column.index[numbers == max_value]
            column[best_algorithms] = "\\tu{"+column[best_algorithms] + '}'
            if critical_dist is not None:
                close_algos = column.index[(numbers-max_value) <= critical_dist]
                column[close_algos] = "\\tb{"+column[close_algos]+'}'
            return column
    if opmaak:
        table = table.apply(opmaak_f, axis = 0)
    print(table.to_latex(escape = False, index = index))

    

# Contamination 5, validation set 10%

datasets_to_use = pd.read_pickle("used_datasets/used_stat_datasets.pkl")
datasets_to_use;

# evaluation_setting_paths = [processed_path/'peak_performance', processed_path]
result_table, critical_distance = \
    generate_result_table(
    ['peak_performance', 'best_average_performance',  'default_performance','statistical_validation_set/p=0.05'],
    reference_for_validation_set = 'statistical_validation_set/p=0.05',
    alpha =0.05, 
    datasets_to_use = datasets_to_use,
)
result_table

small_result_table = result_table.drop(['avg_ap', 'ap_ranks'], axis = 1, level =1)
small_result_table

to_latex(small_result_table)

seperate_tables = (
    small_result_table
    .rename_axis(columns = ['setting', 'metric'])
    .groupby('setting', axis = 1)
)
names = small_result_table.columns.get_level_values(0).unique()
tables = [table.droplevel(0, axis =1).reset_index().sort_values('auc_ranks').reset_index(drop = True) for name, table in seperate_tables]
sorted_small_result_table = pd.concat(tables, axis = 1,keys = names)
sorted_small_result_table

(sorted_small_result_table.loc[:, ('peak_performance', 'avg_auc')] / sorted_small_result_table.loc[:, ('default_performance', 'avg_auc')]).mean()

(sorted_small_result_table.loc[:, ('statistical_validation_set/p=0.05', 'avg_auc')] / sorted_small_result_table.loc[:, ('default_performance', 'avg_auc')]).mean()

to_latex(sorted_small_result_table, index = False, opmaak = False)

print(critical_distance)

# Contamination 5, validation set 20% 

original_contamination, critical_distance = generate_result_table(['peak_performance_v1', 'best_average_performance_v1','default_performance_v1',  'tuned_performance_stat10', 'tuned_performance_stat5'], 'contamination5')
original_contamination

to_latex(original_contamination, critical_distance)

print(critical_distance)

### Contamination 5

contamination5 = compare_algorithms_different_performance_settings('contamination5')
contamination5

to_latex(contamination5)

### all experiments

contamination_all = compare_algorithms_different_performance_settings('all')
contamination_all

to_latex(contamination_all)

# Look at performance distributions

def algorithm_performance_plot(algorithm, contamination_filter, metric = 'auc'): 
    r_path = result_path / f"grid_{algorithm}_v{VERSION}.csv"
    result_df = pd.read_csv(r_path)
    result_df = apply_contamination_filter(contamination_filter, result_df)
    result_df = result_df.assign(dataset_id = lambda x: x.dataset_id +'_' +x.anomaly_fraction.astype('str'))[['algo_name', 'dataset_id', 'auc', 'ap']]
#     result_df = result_df[result_df.dataset_id.str.startswith('ALOI')]
    return alt.Chart(result_df).mark_boxplot().encode(
        x = 'dataset_id:N', 
        y = 'auc:Q'
    ).properties(title = f"{algorithm} data={contamination_filter}")

algorithm_performance_plot('CBLOF', 'contamination5')

algorithm_performance_plot('KNN', 'contamination5')

algorithm_performance_plot('LOF', 'contamination5')

algorithm_performance_plot('IForest', 'contamination5')

algorithm_performance_plot('HBOS', 'contamination5')
